% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jiscode_jaddress.R
\name{jiscode_jaddress}
\alias{jiscode_jaddress}
\title{Search for Address by JIS codes}
\usage{
jiscode_jaddress(x, jis = c("all", "city", "pref"))
}
\arguments{
\item{x}{a a vector. JIS codes.
JIS codes are numbers or character strings composed of half-width numerals, and must have either two, five, or six digits.}

\item{jis}{Choose \code{"all"}, \code{"city"}, or \code{"pref"}, and select the range of addresses to be printed.
\code{"all"} will print both the name of the prefecture and the name of the city.
If you select \code{"all"}, both the name of the prefecture and the name of the city will be output, but if the JIS code is two digits, only the name of the prefecture will be output.
\code{"city"} will output only the name of the city, and \code{"pref"} will output only the name of the prefecture.}
}
\value{
a charactor.
}
\description{
If enter a JIS code, it will output the name of the corresponding prefecture or city.
JIS codes are numbers or character strings composed of half-width numerals, and must have either two, five, or six digits.
}
\details{
The data for the organization code comes from the data published by the Ministry of Internal Affairs and Communications.
Source: Ministry of Internal Affairs and Communications website, National Local Government Code (https://www.soumu.go.jp/denshijiti/code.html) "Prefectural Code and Municipal Code" (updated on May 1, 2019)
The copyright of the data published by the Ministry of Internal Affairs and Communications (MIC) on its website belongs to the Ministry of Internal Affairs and Communications (MIC), but processing and redistribution are permitted within a certain range.
For details, please refer to the web page of the Ministry of Internal Affairs and Communications.
}
