% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaddress_jiscode.R
\name{jaddress_jiscode}
\alias{jaddress_jiscode}
\title{Search for JIS codes by Address}
\usage{
jaddress_jiscode(x, jis = c("city", "pref"), check.digit = FALSE)
}
\arguments{
\item{x}{a vector. address strings.}

\item{jis}{Selecting \code{"city"} will output the group code up to the city, and selecting \code{"pref"} will output the group code up to the prefecture.}

\item{check.digit}{a logical value. If set to \code{TRUE}, the 6th digit check digit will be displayed.}
}
\value{
a charactor.
}
\description{
This function allows you to get the organization code by entering an address that includes the name of the prefecture and city.
The address must be in Japanese, and the name of the prefecture, city, town, or village must be in kanji.
The name of the prefecture must be complete, such as "Akita-ken", and the name of the municipality must end with a string, such as "Akita-shi".
}
\details{
The data for the organization code comes from the data published by the Ministry of Internal Affairs and Communications.
Source: Ministry of Internal Affairs and Communications website, National Local Government Code (https://www.soumu.go.jp/denshijiti/code.html) "Prefectural Code and Municipal Code" (updated on May 1, 2019)
The copyright of the data published by the Ministry of Internal Affairs and Communications (MIC) on its website belongs to the Ministry of Internal Affairs and Communications (MIC), but processing and redistribution are permitted within a certain range.
For details, please refer to the web page of the Ministry of Internal Affairs and Communications.
}
